Create Course_Week 1 and 2
Discussion Forum Week One:
My name is Shane, so the first and last letter of my name (S + E = 19 + 5) give the number  24. The 24th Fibonacci number is 46368. 

Discussion Forum Week Two:
My surname is Healy

The first letter H is number 72

The last letter y is number 121

Fibonacci number 193 is 9663391306290450775010025392525829059713



ord() is a Python function that returns an integer, based on a computing industry standard, to represent the character.

Regards, 
Shane